refactor: extract map layers into seperate files
